Move action_view into location_view
