+++ browser/ui/BUILD.gn
@@ -526,8 +526,6 @@ source_set("ui") {
       "toolbar/brave_location_bar_model_delegate.h",
       "views/brave_actions/brave_actions_container.cc",
       "views/brave_actions/brave_actions_container.h",
-      "views/brave_actions/brave_news_action_view.cc",
-      "views/brave_actions/brave_news_action_view.h",
       "views/brave_actions/brave_rewards_action_view.cc",
       "views/brave_actions/brave_rewards_action_view.h",
       "views/brave_actions/brave_shields_action_view.cc",
@@ -538,6 +536,8 @@ source_set("ui") {
       "views/frame/vertical_tab_strip_widget_delegate_view.h",
       "views/location_bar/brave_location_bar_view.cc",
       "views/location_bar/brave_location_bar_view.h",
+      "views/location_bar/brave_news_location_view.cc",
+      "views/location_bar/brave_news_location_view.h",
       "views/location_bar/brave_star_view.cc",
       "views/location_bar/brave_star_view.h",
       "views/profiles/brave_avatar_toolbar_button.cc",
