+++ browser/ui/views/location_bar/brave_location_bar_view.h
@@ -9,6 +9,7 @@
 #include <vector>
 
 #include "base/gtest_prod_util.h"
+#include "brave/browser/ui/views/location_bar/brave_news_location_view.h"
 #include "brave/components/ipfs/buildflags/buildflags.h"
 #include "brave/components/tor/buildflags/buildflags.h"
 #include "chrome/browser/ui/views/location_bar/location_bar_view.h"
@@ -75,6 +76,7 @@ class BraveLocationBarView : public LocationBarView {
   friend class ::BraveActionsContainerTest;
   friend class ::RewardsBrowserTest;
   BraveActionsContainer* brave_actions_ = nullptr;
+  BraveNewsLocationView* brave_news_location_view = nullptr;
 #if BUILDFLAG(ENABLE_TOR)
   OnionLocationView* onion_location_view_ = nullptr;
 #endif
